Refactor the CompactionController into an abstract version to allow for custom implementations not tied to SSTableReader or reference counting
